+++ packages/pluggableWidgets/barcode-scanner-web/src/hooks/useReader.ts
@@ -8,24 +8,16 @@ import {
     HybridBinarizer,
     NotFoundException,
     Result
-} from "@zxing/library/cjs";
+} from "@zxing/library";
 import { useEventCallback } from "@mendix/widget-plugin-hooks/useEventCallback";
-
-const hints = new Map();
-// RSS_Expanded is not production ready yet.
-const exclusions = new Set([BarcodeFormat.RSS_EXPANDED]);
-// `BarcodeFormat` is a TypeScript enum. Calling `Object.values` on it returns an array of string and ints, we only want the latter.
-const formats = Object.values(BarcodeFormat)
-    .filter((format): format is BarcodeFormat => typeof format !== "string")
-    .filter(format => !exclusions.has(format));
-hints.set(DecodeHintType.POSSIBLE_FORMATS, formats);
+import { BarcodeFormatsType } from "../../typings/BarcodeScannerProps";
 
 const mediaStreamConstraints: MediaStreamConstraints = {
     audio: false,
     video: {
         facingMode: "environment",
-        width: { min: 1280, ideal: 1920, max: 2560 },
-        height: { min: 720, ideal: 1080, max: 1440 }
+        width: { min: 1280, ideal: 2560, max: 2560 },
+        height: { min: 720, ideal: 1440, max: 1440 }
     }
 };
 
@@ -33,64 +25,117 @@ type UseReaderHook = (args: {
     onSuccess?: (data: string) => void;
     onError?: (e: Error) => void;
     useCrop: boolean;
+    barcodeFormats?: BarcodeFormatsType[];
+    useAllFormats: boolean;
+    canvasMiddleRef?: HTMLDivElement;
 }) => RefObject<HTMLVideoElement>;
 
 export const useReader: UseReaderHook = args => {
     const videoRef = useRef<HTMLVideoElement>(null);
     const onSuccess = useEventCallback(args.onSuccess);
     const onError = useEventCallback(args.onError);
-    const scale = 0.3;
-    const stopped = useRef<Boolean>(false);
+    const stopped = useRef<boolean>(false);
+    const reader = useRef<BrowserMultiFormatReader>();
     const checkNotFound = (error: any): boolean => {
         const ifNotFound = error instanceof NotFoundException;
         return ifNotFound && !stopped.current;
     };
 
-    const scanWithCropOnce = (reader: BrowserMultiFormatReader): Promise<Result> => {
-        const cropWidth = videoRef.current!.videoWidth * scale;
-        const captureCanvas = reader.createCaptureCanvas(videoRef.current!);
-        captureCanvas.width = cropWidth;
-        captureCanvas.height = cropWidth;
-        const loop = (resolve: (value: Result) => void, reject: (reason?: Error) => void) => {
-            try {
-                const canvasContext = captureCanvas.getContext("2d");
-                if (canvasContext !== null) {
-                    canvasContext.drawImage(
-                        videoRef.current!,
-                        (videoRef.current!.videoWidth * (1 - scale)) / 2,
-                        (videoRef.current!.videoHeight - cropWidth) / 2,
-                        cropWidth,
-                        cropWidth,
-                        0,
-                        0,
-                        captureCanvas.width,
-                        captureCanvas.width
-                    );
-                    const luminanceSource = new HTMLCanvasElementLuminanceSource(captureCanvas);
-                    const binaryBitmap = new BinaryBitmap(new HybridBinarizer(luminanceSource));
-                    const result = reader.decodeBitmap(binaryBitmap);
-                    resolve(result);
-                }
-            } catch (error) {
-                if (checkNotFound(error)) {
-                    setTimeout(() => loop(resolve, reject), reader.timeBetweenDecodingAttempts);
-                } else {
-                    reject(error);
-                }
+    const loop = (
+        resolve: (value: Result) => void,
+        reject: (reason?: Error) => void,
+        captureCanvas: HTMLCanvasElement,
+        videoCropWidth: number,
+        videoCropHeight: number
+    ): void => {
+        try {
+            const canvasContext = captureCanvas.getContext("2d", { willReadFrequently: true });
+            if (canvasContext !== null && videoRef.current !== null && reader.current !== undefined) {
+                canvasContext.drawImage(
+                    videoRef.current,
+                    (videoRef.current.videoWidth - videoCropWidth) / 2,
+                    (videoRef.current.videoHeight - videoCropHeight) / 2,
+                    videoCropWidth,
+                    videoCropHeight,
+                    0,
+                    0,
+                    videoCropWidth,
+                    videoCropHeight
+                );
+                const luminanceSource = new HTMLCanvasElementLuminanceSource(captureCanvas);
+                const binaryBitmap = new BinaryBitmap(new HybridBinarizer(luminanceSource));
+                const result = reader.current.decodeBitmap(binaryBitmap);
+                resolve(result);
             }
-        };
-        return new Promise(loop);
+        } catch (error) {
+            if (checkNotFound(error)) {
+                setTimeout(() => loop(resolve, reject, captureCanvas, videoCropWidth, videoCropHeight), 50);
+            } else {
+                reject(error);
+            }
+        }
+    };
+
+    const scanWithCropOnce = (reader: BrowserMultiFormatReader): Promise<Result> => {
+        if (videoRef.current && args.canvasMiddleRef) {
+            const aspectRatioClient = videoRef.current.clientWidth / videoRef.current.clientHeight;
+            const aspectRatioVideo = videoRef.current.videoWidth / videoRef.current.videoHeight;
+
+            let videoCropWidth =
+                (args.canvasMiddleRef.clientWidth / videoRef.current.clientWidth) * videoRef.current.videoWidth;
+            let videoCropHeight =
+                (args.canvasMiddleRef.clientHeight / videoRef.current.clientHeight) * videoRef.current.videoHeight;
+
+            if (aspectRatioVideo < aspectRatioClient) {
+                videoCropHeight = videoCropWidth;
+            } else {
+                videoCropWidth = videoCropHeight;
+            }
+
+            const captureCanvas = reader.createCaptureCanvas(videoRef.current);
+            captureCanvas.width = videoCropWidth;
+            captureCanvas.height = videoCropHeight;
+
+            return new Promise((resolve, reject) =>
+                loop(resolve, reject, captureCanvas, videoCropWidth, videoCropHeight)
+            );
+        } else {
+            return Promise.reject(new Error("No video ref or no canvas ref found"));
+        }
     };
 
     useEffect(() => {
-        const reader = new BrowserMultiFormatReader(hints, 500);
-
+        const hints = new Map();
+        let formats: BarcodeFormat[] = [];
+        if (args.useAllFormats) {
+            formats = [
+                BarcodeFormat.UPC_A,
+                BarcodeFormat.UPC_E,
+                BarcodeFormat.EAN_8,
+                BarcodeFormat.EAN_13,
+                BarcodeFormat.CODE_39,
+                BarcodeFormat.CODE_128,
+                BarcodeFormat.ITF,
+                BarcodeFormat.RSS_14,
+                BarcodeFormat.QR_CODE,
+                BarcodeFormat.DATA_MATRIX,
+                BarcodeFormat.AZTEC,
+                BarcodeFormat.PDF_417
+            ];
+        } else {
+            if (args.barcodeFormats) {
+                formats = args.barcodeFormats.map(val => BarcodeFormat[val.barcodeFormat]);
+            }
+        }
+        hints.set(DecodeHintType.POSSIBLE_FORMATS, formats);
+        hints.set(DecodeHintType.ENABLE_CODE_39_EXTENDED_MODE, true);
+        const newReader = new BrowserMultiFormatReader(hints, 500);
+        reader.current = newReader;
         const stop = (): void => {
             stopped.current = true;
-            reader.stopAsyncDecode();
-            reader.reset();
+            newReader.stopAsyncDecode();
+            newReader.reset();
         };
-
         const start = async (): Promise<void> => {
             let stream;
             try {
@@ -102,9 +147,9 @@ export const useReader: UseReaderHook = args => {
                         videoRef.current.autofocus = true;
                         videoRef.current.playsInline = true; // Fix error in Safari
                         await videoRef.current.play();
-                        result = await scanWithCropOnce(reader);
+                        result = await scanWithCropOnce(newReader);
                     } else {
-                        result = await reader.decodeOnceFromStream(stream, videoRef.current);
+                        result = await newReader.decodeOnceFromStream(stream, videoRef.current);
                     }
                     if (!stopped.current) {
                         onSuccess(result.getText());