+++ packages/pluggableWidgets/barcode-scanner-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Added barcode format controls to prevent incorrect scanning
+
 ## [2.3.1] - 2023-09-27
 
 ### Fixed